Refs #20629: Move ServerAttributes implementation
